@@ -16,11 +16,10 @@ classifiers = [
 ]
 keywords = ["weather", "weather-api", "netherlands", "weather-forecast", "pollen", "belgium", "luxembourg", "rain-radar"]
 dependencies = [
-    "aiohttp>=3.11.0",
-    "async-timeout>=4.0.3",
-    "svgwrite>=1.4.3",
+    "aiohttp>=3.11.0,<4.0.0",
+    "svgwrite>=1.4.3,<2.0.0",
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 
 [project.urls]
 Homepage = "https://github.com/jdejaegh/irm-kmi-api"
