@@ -1,9 +1,10 @@
 """Parse pollen info from SVG from IRM KMI api"""
 import logging
 import xml.etree.ElementTree as ET
-from typing import List
+from typing import Dict, List
 
-from .const import POLLEN_LEVEL_TO_COLOR, POLLEN_NAMES
+from .const import POLLEN_LEVEL_TO_COLOR
+from .data import PollenLevel, PollenName
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -21,7 +22,7 @@ class PollenParser:
     ):
         self._xml = xml_string
 
-    def get_pollen_data(self) -> dict:
+    def get_pollen_data(self) -> Dict[PollenName, PollenLevel | None]:
         """
         Parse the SVG and extract the pollen data from the image.
         If an error occurs, return the default value.
@@ -38,11 +39,11 @@ class PollenParser:
 
         elements: List[ET.Element] = self._extract_elements(root)
 
-        pollens = {e.attrib.get('x', None): self._get_elem_text(e).lower()
-                   for e in elements if 'tspan' in e.tag and self._get_elem_text(e) in POLLEN_NAMES}
+        pollens = {e.attrib.get('x', None): self._get_txt(e).lower()
+                   for e in elements if 'tspan' in e.tag and str(self._get_txt(e)).lower() in PollenName}
 
-        pollen_levels = {e.attrib.get('x', None): POLLEN_LEVEL_TO_COLOR[self._get_elem_text(e)]
-                         for e in elements if 'tspan' in e.tag and self._get_elem_text(e) in POLLEN_LEVEL_TO_COLOR}
+        pollen_levels = {e.attrib.get('x', None): POLLEN_LEVEL_TO_COLOR[self._get_txt(e)]
+                         for e in elements if 'tspan' in e.tag and self._get_txt(e) in POLLEN_LEVEL_TO_COLOR}
 
         level_dots = {e.attrib.get('cx', None) for e in elements if 'circle' in e.tag}
 
@@ -50,13 +51,18 @@ class PollenParser:
         # As of January 2025, the text is always 'active' and the dot shows the real level
         # If text says 'active', check the dot; else trust the text
         for position, pollen in pollens.items():
+            # Check if pollen is a known one
+            try:
+                pollen: PollenName = PollenName(pollen)
+            except ValueError:
+                _LOGGER.warning(f'Unknown pollen name {pollen}')
+                continue
             # Determine pollen level based on text
             if position is not None and position in pollen_levels:
                 pollen_data[pollen] = pollen_levels[position]
-                _LOGGER.debug(f"{pollen} is {pollen_data[pollen]} according to text")
-
+                _LOGGER.debug(f"{pollen.value} is {pollen_data[pollen]} according to text")
             # If text is 'active' or if there is no text, check the dot as a fallback
-            if pollen_data[pollen] not in {'none', 'active'}:
+            if pollen_data[pollen] not in {PollenLevel.NONE, PollenLevel.ACTIVE}:
                 _LOGGER.debug(f"{pollen} trusting text")
             else:
                 for dot in level_dots:
@@ -66,35 +72,35 @@ class PollenParser:
                         pass
                     else:
                         if 24 <= relative_x_position <= 34:
-                            pollen_data[pollen] = 'green'
+                            pollen_data[pollen] = PollenLevel.GREEN
                         elif 13 <= relative_x_position <= 23:
-                            pollen_data[pollen] = 'yellow'
+                            pollen_data[pollen] = PollenLevel.YELLOW
                         elif -5 <= relative_x_position <= 5:
-                            pollen_data[pollen] = 'orange'
+                            pollen_data[pollen] = PollenLevel.ORANGE
                         elif -23 <= relative_x_position <= -13:
-                            pollen_data[pollen] = 'red'
+                            pollen_data[pollen] = PollenLevel.RED
                         elif -34 <= relative_x_position <= -24:
-                            pollen_data[pollen] = 'purple'
+                            pollen_data[pollen] = PollenLevel.PURPLE
 
-                _LOGGER.debug(f"{pollen} is {pollen_data[pollen]} according to dot")
+                _LOGGER.debug(f"{pollen.value} is {pollen_data[pollen]} according to dot")
 
         _LOGGER.debug(f"Pollen data: {pollen_data}")
         return pollen_data
 
     @staticmethod
-    def get_default_data() -> dict:
+    def get_default_data() -> Dict[PollenName, PollenLevel | None]:
         """Return all the known pollen with 'none' value"""
-        return {k.lower(): 'none' for k in POLLEN_NAMES}
+        return {k: PollenLevel.NONE for k in PollenName}
 
     @staticmethod
-    def get_unavailable_data() -> dict:
+    def get_unavailable_data() -> Dict[PollenName, PollenLevel | None]:
         """Return all the known pollen with None value"""
-        return {k.lower(): None for k in POLLEN_NAMES}
+        return {k: None for k in PollenName}
 
     @staticmethod
-    def get_option_values() -> List[str]:
+    def get_option_values() -> List[PollenLevel]:
         """List all the values that the pollen can have"""
-        return list(POLLEN_LEVEL_TO_COLOR.values()) + ['none']
+        return list(POLLEN_LEVEL_TO_COLOR.values()) + [PollenLevel.NONE]
 
     @staticmethod
     def _extract_elements(root) -> List[ET.Element]:
@@ -106,7 +112,7 @@ class PollenParser:
         return elements
 
     @staticmethod
-    def _get_elem_text(e) -> str | None:
+    def _get_txt(e) -> str | None:
         if e.text is not None:
             return e.text.strip()
         return None
