@@ -3,13 +3,12 @@ import datetime
 import json
 from datetime import datetime as dt
 from datetime import timedelta
-from unittest.mock import MagicMock, AsyncMock
+from unittest.mock import AsyncMock, MagicMock
 from zoneinfo import ZoneInfo
 
-from irm_kmi_api.api import IrmKmiApiClientHa
-from irm_kmi_api.const import OPTION_STYLE_SATELLITE
-from irm_kmi_api.data import AnimationFrameData, RadarAnimationData
-from irm_kmi_api.rain_graph import RainGraph
+from irm_kmi_api import IrmKmiApiClientHa
+from irm_kmi_api import AnimationFrameData, RadarAnimationData, RadarStyle
+from irm_kmi_api import RainGraph
 from tests.conftest import load_fixture
 
 
@@ -45,7 +44,7 @@ async def test_svg_frame_setup():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     await rain_graph._draw_svg_frame()
@@ -65,7 +64,7 @@ def test_svg_hint():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._write_hint()
@@ -81,7 +80,7 @@ def test_svg_time_bars():
         tz = datetime.UTC,
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._draw_hour_bars()
@@ -100,7 +99,7 @@ def test_draw_chances_path():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._draw_chances_path()
@@ -118,7 +117,7 @@ def test_draw_data_line():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._draw_data_line()
@@ -136,7 +135,7 @@ async def test_insert_background():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     await rain_graph._insert_background()
@@ -159,7 +158,7 @@ def test_draw_current_frame_line_moving():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._draw_current_fame_line()
@@ -187,7 +186,7 @@ def test_draw_current_frame_line_index():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._draw_current_fame_line(0)
@@ -216,7 +215,7 @@ def test_draw_description_text():
         tz=datetime.UTC,
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._draw_description_text()
@@ -243,7 +242,7 @@ def test_draw_cloud_layer():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._insert_cloud_layer()
@@ -263,7 +262,7 @@ async def test_draw_location_layer():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     await rain_graph._draw_location()
@@ -281,7 +280,7 @@ def test_get_animation_data():
 
     tz = ZoneInfo('Europe/Brussels')
     lang = 'en'
-    style = OPTION_STYLE_SATELLITE
+    style = RadarStyle.OPTION_STYLE_SATELLITE
     dark_mode = False
 
     api._api_data = json.loads(load_fixture("forecast.json"))
@@ -306,7 +305,7 @@ async def test_download_single_cloud():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._api_client = MagicMock()
@@ -324,7 +323,7 @@ async def test_download_many_clouds():
     rain_graph = RainGraph(
         animation_data=data,
         country='BE',
-        style='STD',
+        style=RadarStyle.OPTION_STYLE_STD,
     )
 
     rain_graph._api_client = MagicMock()
@@ -339,11 +338,11 @@ def test_can_build_rain_graph_with_empty_sequence():
 
     RainGraph(
         RadarAnimationData(sequence=None),
-        'en', 'style'
+        'BE', RadarStyle.OPTION_STYLE_STD
     )
 
     RainGraph(
         RadarAnimationData(sequence=[]),
-        'en', 'style'
+        'BE', RadarStyle.OPTION_STYLE_STD
     )
 
