@@ -2,6 +2,7 @@ from datetime import datetime
 
 from freezegun import freeze_time
 
+from irm_kmi_api import WarningType
 from tests.conftest import get_api_with_data
 
 
@@ -18,7 +19,7 @@ async def test_warning_data() -> None:
     assert first.get('starts_at').replace(tzinfo=None) < datetime.now()
     assert first.get('ends_at').replace(tzinfo=None) > datetime.now()
 
-    assert first.get('slug') == 'fog'
+    assert first.get('slug') == WarningType.FOG
     assert first.get('friendly_name') == 'Fog'
     assert first.get('id') == 7
     assert first.get('level') == 1
